pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["judoree <135936771+judoree@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.14.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.221.1"}
strawberry-graphql-django = "^0.35.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
